requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asttokens==2.4.1
    # via stack-data
black==24.4.2
    # via -r requirements.in
build==1.2.1
    # via pip-tools
click==8.1.7
    # via
    #   black
    #   pip-tools
decorator==5.1.1
    # via ipython
executing==2.0.1
    # via stack-data
flake8==7.0.0
    # via -r requirements.in
ipython==8.24.0
    # via -r requirements.in
jedi==0.19.1
    # via ipython
matplotlib-inline==0.1.7
    # via ipython
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
packaging==24.0
    # via
    #   black
    #   build
parso==0.8.4
    # via jedi
pathspec==0.12.1
    # via black
pexpect==4.9.0
    # via ipython
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.2.2
    # via black
prompt-toolkit==3.0.44
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.11.1
    # via flake8
pyflakes==3.2.0
    # via flake8
pygments==2.18.0
    # via ipython
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
python-decouple==3.8
    # via -r requirements.in
six==1.16.0
    # via
    #   asttokens
    #   thrift
stack-data==0.6.3
    # via ipython
thrift==0.20.0
    # via -r requirements.in
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
wcwidth==0.2.13
    # via prompt-toolkit
wheel==0.43.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
